Rename shlink-api-proxy and save-tag-colors route files
